+++ src/main/java/pl/rscorporation/bookstoreapi/manager/BookService.java
@@ -3,43 +3,70 @@ package pl.rscorporation.bookstoreapi.manager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import pl.rscorporation.bookstoreapi.dao.AuthorRepository;
 import pl.rscorporation.bookstoreapi.dao.BookRepository;
+import pl.rscorporation.bookstoreapi.dao.dto.BookReadDTO;
+import pl.rscorporation.bookstoreapi.dao.dto.BookWriteDTO;
+import pl.rscorporation.bookstoreapi.dao.models.Author;
 import pl.rscorporation.bookstoreapi.dao.models.Book;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Optional;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class BookService {
 
     private BookRepository bookRepository;
+    private AuthorRepository authorRepository;
 
-    @Autowired
-    public BookService(BookRepository bookRepository){
+    public BookService(BookRepository bookRepository, AuthorRepository authorRepository){
         this.bookRepository = bookRepository;
+        this.authorRepository = authorRepository;
     }
 
-    @EventListener(ApplicationReadyEvent.class)
-    public void fillDB(){
-        save(new Book(1L, "1111111111", "tytuł", 11L, BigDecimal.valueOf(24.5), "PLN", LocalDate.of(2002, 1, 2)));
-        save(new Book(2L, "2222222222", "tytuł2", 11L, BigDecimal.valueOf(19.9), "PLN", LocalDate.of(2020, 2, 28)));
+//    public List<Book> findBookByAuthorId(Long id){
+//        return authorRepository.findById(id)
+//                .map(author -> bookRepository.findByAuthorId(author.getId()))
+//                .orElseThrow(() -> new IllegalArgumentException("Author with given id not exists"));
+//    }
+
+    public BookReadDTO findBookById(Long id){
+        Book toReturn = bookRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Book with this id not exists"));
+        return new BookReadDTO(toReturn);
     }
 
-    public Optional<Book> findById(Long id){
-        return bookRepository.findById(id);
+    public List<BookReadDTO> findAll(){
+        return bookRepository.findAll().stream()
+                .map(book -> new BookReadDTO(book))
+                .collect(Collectors.toList());
     }
 
-    public Iterable<Book> findAll(){
-        return bookRepository.findAll();
+    public List<BookReadDTO> findAll(Pageable page){
+        return bookRepository.findAll(page)
+                .stream()
+                .map(BookReadDTO::new)
+                .collect(Collectors.toList());
     }
 
-    public Book save(Book book){
-        return bookRepository.save(book);
+    public BookReadDTO addBook(BookWriteDTO book){
+        Author bookAuthor = authorRepository.findById(book.getAuthorId())
+                .orElseThrow(() -> new IllegalArgumentException("Author with given id not exists insert correct book author"));
+
+        Book saved = bookRepository.save(book.toBook(bookAuthor));
+        return new BookReadDTO(saved);
     }
 
-    public void deleteById(Long id){
+
+    public void deleteBookById(Long id)
+    {
+        if(!bookRepository.existsById(id))
+            throw new IllegalArgumentException("Book with given id not exists");
         bookRepository.deleteById(id);
     }
 }
